# Optimize `#pluralize`