Several different optimization, see each commit message for details.

Before:

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Calculating -------------------------------------
             regular    252.175k (± 0.3%) i/s    (3.97 μs/i) -      1.275M in   5.054643s
           irregular    851.502k (± 1.0%) i/s    (1.17 μs/i) -      4.316M in   5.069125s
         uncountable      1.487M (± 0.7%) i/s  (672.56 ns/i) -      7.578M in   5.097231s
```

After:

```
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Calculating -------------------------------------
             regular    298.382k (± 0.5%) i/s    (3.35 μs/i) -      1.515M in   5.076864s
           irregular      1.943M (± 0.4%) i/s  (514.77 ns/i) -      9.758M in   5.023349s
         uncountable      6.008M (± 0.4%) i/s  (166.45 ns/i) -     30.366M in   5.054350s
```

Benchmark:

```ruby
# frozen_string_literal: true
require "bundler/inline"

gemfile do
  gem "rails", path: "."
  gem "benchmark-ips"
end

Benchmark.ips do |x|
  x.report("regular") { ActiveSupport::Inflector.pluralize("test") }
  x.report("irregular") { ActiveSupport::Inflector.pluralize("zombie") }
  x.report("uncountable") { ActiveSupport::Inflector.pluralize("sheep") }
end
```